requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
openai==1.35.13
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
requests==2.31.0
pytest==8.2.0
